Brave, Monster クラスの分割
